implement whatIsNameInDecade() and tests.
All the decade answers given in the course video are wrong, because Jennifer is the 17th most popular name in 1994, not the 21st.  But even if the name was 21st, the decade files have different names in the 21st row than what the video showed.  Code here matches the data files (I examined each by hand to determine the expected result before running the test.)