minor test case update, minor doc update
After a LOT of pain finally got this running in Intellij.  Found a doc error.  Added a test case for full genderPronoun() coverage.  Rest of modules are fully covered, expect for the getCount exception handler, because I didn't try using a data file with a bad count entry.  I can live with that.